Revert all client changes.

+++ clients/unshaded/src/main/java/tachyon/client/UfsUtils.java
@@ -162,7 +162,7 @@ public class UfsUtils {
           LOG.info("File " + tfsPath + " already exists in Tachyon.");
           continue;
         }
-        long fileId = tfs.createFile(tfsPath, ufsPath);
+        int fileId = tfs.createFile(tfsPath, ufsPath);
         if (fileId == -1) {
           LOG.warn("Failed to create tachyon file: " + tfsPath);
         } else {
